PDT1/hito2/requirements.txt
# Core
langchain==0.3.27
llama-index==0.14.5
langchain_community
google-genai
google-generativeai
chromadb
langchain-chroma
markdown2
weasyprint

# Document loaders
PyPDF2==3.0.1
unstructured==0.18.15
beautifulsoup4==4.14.2
lxml==4.9.3
html5lib==1.1
python-docx==1.2.0
python-pptx==1.0.2

# Utilities
tqdm==4.66.1
requests==2.31.0
PyYAML==6.0.3
python-dotenv
